Move tear-down service
